restructure to separate root and version
